requirements.txt
chardet
guesslang
matplotlib
pandas
PyYAML
requests
